chore: refactor rabbitmq connector

+++ include/app_config_loader.h
@@ -6,7 +6,7 @@
 class AppConfigLoader
 {
 public:
-    static size_t get_env_or_default(const std::string &env_var, const size_t &default_value);
+    static int get_env_or_default(const std::string &env_var, const size_t &default_value);
     static std::string get_env_or_default(const std::string &env_var, const std::string &default_value);
     static std::string get_env_required(const std::string &env_var);
 };
